Schema edited
Drops tables every restart to facilitate easier testing.

Either need to edit JSON data injected for attractions to not be null in return for Get request. But, also need to fix the API side of the application with ratings still seemingly not being added and avg rating still now showing in the get requests.